requirements.txt
customtkinter
openai
anthropic
google-generativeai
groq
llama-cpp-python
sqlalchemy
pyperclip
pydantic
keyring
psutil
GPUtil
ollama
